V0.1.5
Exp system almost working correctly.
inconsistency between exp chacker and xp_math
xp_math might be the culprit, it does not account for full count, only for current level and gained exp in the moment.  but why shud this be an issue?  is it out of sync?  this only happens below level 20  and it reverts consistently to 15 for some reason